requirements.txt
sockslib
pycryptodome